bring back 'api_version' and monkeypatching
There is no good enough reason not to do the same things as fdb
ie. monkeypatching and making a call to api_version mandatory before
@transactional is used.

The alternative, I tried, is not optimal and leads the user of
the library to non-pythonic use of a third party library. That is
it required the user of found to import the version of the library
they want to use in 'foo.fdb' module and use that reference to the
library in their project.

No monkeypatching makes impossible to create distributable layers.

The main advantage of not relying on monkeypatching was good IDE
support. That being said, the API is small enough and emacs smart
enough to deal with the monkeypatch.

Fix #1