feature: hooks and states
Hooks _should_ allow to prepare then validate data before commiting.

States allows to share data inside a transaction without having to
pass the arguments all around the place e.g. user_version in
versionstamp.  As part of the original vnstore it was also used to
keep around the information that a commit was already started during
the transaction. In effect, that the transaction was a commit.

Hooks and states were added late during the SRFI-167 process.

ref: https://srfi.schemers.org/srfi-167/